non so cosa scrivere

+++ inc/shared_memory.h
@@ -24,6 +24,6 @@ void free_shared_memory(void *ptr_sh);
 void remove_shared_memory(int shmid);
 int shmem_add(msg_t **dest, msg_t *src);
 int there_is_message(msg_t **shmem);
-int is_empty(msg_t **shmem, int index);
+int is_empty(msg_t **shmem, size_t index);
 
 #endif
